Tested all scenarios of dca except deposit native having issue
